% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archnetan-package.R
\docType{package}
\name{archnetan-package}
\alias{archnetan}
\alias{archnetan-package}
\title{archnetan: Network Similarity Analysis for Archaeology}
\description{
A system to enable network similarity/community analysis between different araheological deposits (e.g. obsidian tool deposits in New Zealand).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nesaking100/archnetan}
  \item Report bugs at \url{https://github.com/nesaking100/archnetan/issues}
}

}
\author{
\strong{Maintainer}: Joshua Looker \email{josh@looker.co.nz}

}
\keyword{internal}
