% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_loading.R
\name{archnetan_example}
\alias{archnetan_example}
\title{Get path to example csvs (heavily inspired by readr::readr_example)}
\usage{
archnetan_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
archnetan includes csvs used during the analysis of corresponding publications included in
inst/examples which this function helps make easier to access
}
\examples{
archnetan_example()
archnetan_example("originalDataset.csv")
}
