% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_similarity.R
\name{calc_br_sim}
\alias{calc_br_sim}
\title{Calculates the similarities of a bipartite network using the Brainerd-Robinson coefficient}
\usage{
calc_br_sim(df_graph)
}
\arguments{
\item{df_graph}{A dataframe of edge weights of a bipartite network}
}
\value{
An output list containing three components:
\item{col1_brsim}{Tibble containing the Brainerd-Robinson coefficients of first node type/columns}
\item{col2_brsim}{Tibble containing the Brainerd-Robinson coefficients of second node type/columns}
\item{col1_brmat}{Tibble containing the Brainerd-Robinson coefficient matrix for the first column}
\item{col2_brmat}{Tibble containing the Brainerd-Robinson coefficient matrix for the second column}
}
\description{
Calculates the similarities of a bipartite network using the Brainerd-Robinson coefficient
}
\examples{
#calc_br_sim(df_graph_main)
}
