% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_community.R
\name{calc_com_louv}
\alias{calc_com_louv}
\title{Apply the Louvain Community Algorithm}
\usage{
calc_com_louv(df_sim)
}
\arguments{
\item{df_sim}{A dataframe or tibble containing similarity scores}
}
\value{
An output list containing two components:
\item{community}{igraph::communities object of the Louvain communities}
\item{G}{igraph::graph object of the graph}
}
\description{
Apply the Louvain Community Algorithm
}
\examples{
#calc_com_louv(df_similarities)
}
