% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_similarity.R
\name{calc_cos_sim}
\alias{calc_cos_sim}
\title{Calculates the cosine similarities of a bipartite network}
\usage{
calc_cos_sim(df_graph)
}
\arguments{
\item{df_graph}{A dataframe of edge weights of a bipartite network}
}
\value{
An output list containing two components:
\item{col1_cossim}{Tibble containing the cosine similarities of first node type}
\item{col2_cossim}{Tibble containing the cosine similarities of the second node type}
}
\description{
Calculates the cosine similarities of a bipartite network
}
\examples{
#calc_cos_sim(df_graph_main)
}
