% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_similarity.R
\name{calc_rca_sim}
\alias{calc_rca_sim}
\title{Calculates the cosine similarities of a bipartite network}
\usage{
calc_rca_sim(df_graph, type1 = "Source", type2 = "POI")
}
\arguments{
\item{df_graph}{A dataframe of edge weights of a bipartite network}

\item{type1}{A string detailing the name of the first node type}

\item{type2}{A string detailing the name of the second node type}
}
\value{
An output list containing three components:
\item{col1_cossim}{Tibble containing the cosine similarities of first node type/columns}
\item{col2_cossim}{Tibble containing the cosine similarities of the second node type/rows}
\item{RCA_full}{Tibble containing the unnormalised RCA values of both types}
}
\description{
Calculates the cosine similarities of a bipartite network
}
\examples{
#calc_cos_sim(df_graph_main, "Source", "PoI")
}
