% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition_robustness_cos.R
\name{cos_communities_single}
\alias{cos_communities_single}
\title{Adds noise (in the form of full data points) and then create cosine-similarity communities from an inputted dataset}
\usage{
cos_communities_single(df, col1, col2, proportion = 0.1, numReps = 100)
}
\arguments{
\item{df}{Dataframe original frequency graph}

\item{col1}{A string of the first column used to make graphs}

\item{col2}{A string of the second column used to make graphs}

\item{proportion}{A double corresponding to the proportion of the original dataset's size to add as noise}

\item{numReps}{Number of runs noise addition runs to complete}
}
\value{
An output list containing nine components:
\item{comm1}{Tibble containing the assigned communities of column 1 in each run}
\item{comm2}{Tibble containing the assigned communities of column 2 in each run}
\item{hamming1}{Vector containing the Hamming distance to the original column 1 communities in each run}
\item{hamming2}{Vector containing the Hamming distance to the original column 1 communities in each run}
\item{col1Changes}{Tibble of binary values for whether a column 1 node changes communities in each run}
\item{col2Changes}{Tibble of binary values for whether a column 1 node changes communities in each run}
\item{df}{List of lists corresponding to the df_graph like tibbles produced in each bootstrap}
}
\description{
Adds noise (in the form of full data points) and then create cosine-similarity communities from an inputted dataset
}
\examples{
\dontrun{
cos_communities_single(df_graph, "Source", "Site", proportion=0.2, numReps=50)
}
}
