% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition_robustness_cos.R
\name{create_added_noise_cos_communities}
\alias{create_added_noise_cos_communities}
\title{Runs cos_communities_single or cos_communities multiple over a range of proportions and aggregates Hamming distances}
\usage{
create_added_noise_cos_communities(
  df,
  col1,
  col2,
  proportion = 0.1,
  numReps = 100,
  method = "single",
  propSave = 0.1
)
}
\arguments{
\item{df}{Dataframe original frequency graph}

\item{col1}{A string of the first column used to make graphs}

\item{col2}{A string of the second column used to make graphs}

\item{proportion}{A double corresponding to the proportion of the original dataset's size to add as noise}

\item{numReps}{Number of runs noise addition runs to complete}

\item{method}{A string determining the method of drawing added noise data points: single for full artifact,
multi for drawing each artifact variably separately}

\item{propSave}{Proportion of noise added for which to save the cos_communities_single/multi output for}
}
\value{
An output list containing three components:
\item{col1Stats}{Tibble containing the mean and standard deviation in column 1 in Hamming distance for each proportion}
\item{col1Stats}{Tibble containing the mean and standard deviation in column 2 in Hamming distance for each proportion}
\item{propSave}{List in same structure as outputs from cos_communities_single/multi for chosen proportion value}
}
\description{
Runs cos_communities_single or cos_communities multiple over a range of proportions and aggregates Hamming distances
}
\examples{
\dontrun{
create_added_noise_cos_communities(df_graph, "Source", "Site",
proportion=c(0.2,0.3,0.4,0.5), numReps=50, method="multi", propSave=0.5)
}
}
