% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_robustness.R
\name{create_bootstrap_communities}
\alias{create_bootstrap_communities}
\title{Bootstraps communities from unweighted dataset}
\usage{
create_bootstrap_communities(df, col1, col2, nBoot = 100)
}
\arguments{
\item{df}{A tibble of the unweighted dataset}

\item{col1}{First column of network}

\item{col2}{Second column of network}

\item{nBoot}{Number of bootstrap replications to create}
}
\value{
An output list containing three components:
\item{comm1}{Tibble containing bootstrapped communities of col1}
\item{comm2}{Tibble containing bootstrapped communities of col2}
\item{df}{Nested list of bootstrap replicant weighted dataframes}
}
\description{
Bootstraps communities from unweighted dataset
}
\examples{
#create_boostrap_communities(base_df, "Source", "Combined")
}
