% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_building.R
\name{create_graph_single_file}
\alias{create_graph_single_file}
\title{Creates a bipartite network from a single input file}
\usage{
create_graph_single_file(input_string, col1, col2)
}
\arguments{
\item{input_string}{A path to the input file (assumed to be a csv)}

\item{col1}{A string of the first column to create a network from}

\item{col2}{A string of the second column to create a network from}
}
\value{
An output list containing two components:
\item{df}{Tibble containing the edge weights of the bipartite network}
\item{graph}{igraph::graph object, the bipartite network}
}
\description{
Creates a bipartite network from a single input file
}
\examples{
#create_graph_single_file("trial.csv","Source","POI")
}
