% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_building.R
\name{create_graphs_two_files}
\alias{create_graphs_two_files}
\title{Creates one or two bipartite networks from two input files}
\usage{
create_graphs_two_files(
  input_string1,
  input_string2,
  col1,
  col2,
  col3 = NULL,
  col4 = NULL,
  join_string
)
}
\arguments{
\item{input_string1}{A path to the first input file (assumed to be a csv)}

\item{input_string2}{A path to the second input file (assumed to be a csv)}

\item{col1}{First column of first network}

\item{col2}{Second column of first network}

\item{col3}{First column of second network (Optional)}

\item{col4}{Second column of second network (Optional)}

\item{join_string}{Column to merge the two input csvs on}
}
\value{
An output list containing five components:
\item{df_12}{Tibble containing the edge weights of the first bipartite network}
\item{graph_12}{igraph::graph object, the first bipartite network}
\item{df_34}{Tibble containing the edge weights of the second bipartite network - requires col3,col4}
\item{graph_34}{igraph::graph object, the second bipartite network - requires col3,col4}
\item{df_full}{Raw, merged dataframe of the two input files}
}
\description{
Creates one or two bipartite networks from two input files
}
\examples{
#create_graphs_two_files("trial.csv", "trial2.csv", "Source",
# "Litho", "Meta_Assemblage", "Meta_Litho", "Source")
}
