% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_robustness.R
\name{create_permutation_communities}
\alias{create_permutation_communities}
\title{Creates communities by switching edges within a weighted edgelist}
\usage{
create_permutation_communities(df, col1, col2)
}
\arguments{
\item{df}{A tibble of the weighted edgelist}

\item{col1}{First column of network}

\item{col2}{Second column of network}
}
\value{
An output list containing three components:
\item{comm1}{Tibble containing permuted communities of col1}
\item{comm2}{Tibble containing permuted communities of col2}
\item{df}{Nested list of permuted weighted dataframes}
}
\description{
Creates communities by switching edges within a weighted edgelist
}
\examples{
#create_permutation_communities(base_df, "Source", "Combined")
}
