% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_building.R
\name{create_weighted_graph_df}
\alias{create_weighted_graph_df}
\title{Creates a bipartite network from an inputted dataframe with weighted frequencies}
\usage{
create_weighted_graph_df(df, col1, col2, colWt)
}
\arguments{
\item{df}{A dataframe to create the network from (assumed to be a tibble)}

\item{col1}{First column of network}

\item{col2}{Second column network}

\item{colWt}{Column of weights in df}
}
\value{
An output list containing two components:
\item{df}{Tibble containing the edge weights of the bipartite network}
\item{graph}{igraph::graph object, the bipartite network}
}
\description{
Creates a bipartite network from an inputted dataframe with weighted frequencies
}
\examples{
#create_weighted_graph_df(df_full, "Source", "Combined", "Weight (g)")
}
