% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obsidian_full_df}
\alias{obsidian_full_df}
\title{obsidian_full_df}
\format{
A data frame with seven variables:
\describe{
\item{\code{Study Site}}{Assemblage location where the obsidian artifact was found}
\item{\code{Source}}{Geographical origin of the obsidian artifact}
\item{\code{Period Start}}{Start of the likely time period where the artifact was deposited}
\item{\code{Period End}}{End of the likely time period where the artifact was deposited}
\item{\code{Probable Period}}{Whether the likely time period is 'early' or 'late'}
\item{\code{Weight (g)}}{Weight of the artifact in grams}
\item{\code{Reference}}{Citation for the original reference of the artifact}
}
}
\source{
\url{https://doi.org/10.5334/joad.52}
}
\usage{
obsidian_full_df
}
\description{
Dataframe of New Zealand obsidian data collected by collaborators
and package co-authors
}
\keyword{datasets}
