% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obsidian_graph_df}
\alias{obsidian_graph_df}
\title{obsidian_graph_df}
\format{
A data frame with three variables:
\describe{
\item{\code{Study Site}}{Assemblage location where the obsidian artifact was found}
\item{\code{Source}}{Geographical origin of the obsidian artifact}
\item{\code{weight}}{Edgeweight in the corresponding bipartite graph of Study Site and Source}
}
}
\source{
\url{https://doi.org/10.5334/joad.52}
}
\usage{
obsidian_graph_df
}
\description{
Frequency weighted bipartite graph of New Zealand obsidian data collected by collaborators
and package co-authors
}
\keyword{datasets}
