% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_similarity.R
\name{calc_rca_sim}
\alias{calc_rca_sim}
\title{Calculates the cosine similarities of a bipartite network using RCA normalisation}
\usage{
calc_rca_sim(df_graph)
}
\arguments{
\item{df_graph}{A dataframe of edge weights of a bipartite network}
}
\value{
An output list containing three components:
\item{col1_cossim}{Tibble containing the cosine similarities of first node type/column after RCA normalisation}
\item{col2_cossim}{Tibble containing the cosine similarities of the second node type/column after RCA normalisation}
\item{RCA_full}{Tibble containing the RCA values of both columns/node types}
}
\description{
Calculates the cosine similarities of a bipartite network using RCA normalisation
}
\examples{
#calc_rca_sim(df_graph_main)
}
